refactor: split prometheus_recording into per-metric-family mixin package (#2050)
